+++ src/javascript/icons.js
@@ -1,0 +1,53 @@
+const icons = {
+	split_screen : `
+		<svg  xmlns="http://www.w3.org/2000/svg" width="20" height="14" viewBox="0 0 14 14">
+		  <g  id="Grupo_6" data-name="Grupo 6" transform="translate(-472 -381)">
+		    <g  class="bbo" id="Rectángulo_34" data-name="Rectángulo 34" transform="translate(472 381)" fill="transparent"  stroke-width="1.5">
+		      <rect width="14" height="14" rx="2" stroke="none"/>
+		      <rect x="0.75" y="0.75" width="12.5" height="12.5" rx="1.25" fill="none"/>
+		    </g>
+		    <path class="aba" id="Trazado_27" data-name="Trazado 27" d="M5,0A2,2,0,0,1,7,2V12a2,2,0,0,1-2,2H2a11.389,11.389,0,0,1-2-.125V2S-.011.206,0,0Z" transform="translate(479 381)" />
+		    <g class="bba" id="Rectángulo_36" data-name="Rectángulo 36" transform="translate(482 386)"  fill="white"  stroke-width="1">
+		      <rect width="1" height="4" stroke="none"/>
+		      <rect x="0.5" y="0.5" height="3" fill="none"/>
+		    </g>
+		    <g class="bba" id="Rectángulo_37" data-name="Rectángulo 37" transform="translate(484.5 387.5) rotate(90)" fill="white" stroke-width="1">
+		      <rect width="1" height="4" stroke="none"/>
+		      <rect x="0.5" y="0.5" height="3" fill="none"/>
+		    </g>
+		  </g>
+		</svg>
+	`,
+	remove_screen:`
+		<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="20" height="14" viewBox="0 0 14 14">
+	  <g id="remove_screen" >
+	    <g class="bbo" id="_2" data-name="2" fill="none"  stroke-width="1.5">
+	      <rect width="14" height="14" rx="2" stroke="none"/>
+	      <rect x="0.75" y="0.75" width="12.5" height="12.5" rx="1.25" fill="none"/>
+	    </g>
+	    <path class="aba" id="_1" data-name="1" d="M4.756,0a1.953,1.953,0,0,1,1.9,2V12a1.953,1.953,0,0,1-1.9,2H1.9A10.318,10.318,0,0,1,0,13.875V2S-.01.206,0,0Z" transform="translate(7.341)" fill="#707070"/>
+	    <g class="aba" id="_0" data-name="0" transform="translate(6.5 6.5) rotate(90)" stroke-width="1">
+	      <rect width="1" height="4" stroke="none"/>
+	      <rect x="0.5" y="0.5" height="3" fill="none"/>
+	    </g>
+	  </g>
+	</svg>
+	`,
+	settings:`
+		<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="20" height="14" viewBox="0 0 14 14">
+	  <g id="Componente_16_1" data-name="Componente 16 – 1" fill="transparent" >
+	    <g class="bbo" id="Elipse_3" data-name="Elipse 3"  stroke-width="1.5">
+	      <circle  cx="7" cy="7" r="7" />
+	      <circle  cx="7" cy="7" r="6.25" />
+	    </g>
+	    <circle class="bbo" id="Elipse_4" data-name="Elipse 4" cx="3" cy="3" r="3" transform="translate(4 4)" stroke-dasharray="1 2"  stroke-linecap="round" stroke-linejoin="round" stroke-width="1" />
+	  </g>
+	</svg>
+	`,
+	empty : `
+		<svg xmlns="http://www.w3.org/2000/svg" width="20" height="15" viewBox="0 0 14 14">
+		  
+		</svg>
+	`
+
+}